Latitud y Longitud para productos ARREGLADO

+++ src/main/java/com/gameexchange/gameexchange/domain/Producto.java
@@ -38,9 +38,9 @@ public class Producto implements Serializable {
     private Long videojuego;
 
     @Column(name = "latitud")
-    private Long latitud;
+    private Double latitud;
     @Column(name = "longitud")
-    private Long longitud;
+    private Double longitud;
 
     @ManyToOne
     private User usuario;
@@ -189,19 +189,19 @@ public class Producto implements Serializable {
         this.ventas = ventas;
     }
 
-    public Long getLatitud() {
+    public Double getLatitud() {
         return latitud;
     }
 
-    public void setLatitud(Long latitud) {
+    public void setLatitud(Double latitud) {
         this.latitud = latitud;
     }
 
-    public Long getLongitud() {
+    public Double getLongitud() {
         return longitud;
     }
 
-    public void setLongitud(Long longitud) {
+    public void setLongitud(Double longitud) {
         this.longitud = longitud;
     }
